Add paddle side collisions
Add collisions to the paddle sides (top and bottom) so that the ball
doesn't ignore the paddle and pass through it if it misses its front.
The sides of the paddle now cause the ball to bounce away but still
continue towards the side of the display behind the paddle.