Add spin to ball
Add spin to ball, which is caused by - and affected by - the speed of
the paddle when the ball collides with one. Adjust paddle acceleration
and deceleration to make player paddle smoother and less likely to be
considered still when colliding with the ball.
The keys "<" and ">" can be used to increase left or right spin
respectively.