Tune ball spin
Set maximum spin for ball. Make sure that paddle's speed is set to zero
when paddle hits the edge of the display. 'Deaden' the ball's spin when
it hits a paddle or a wall (top or bottom of display). Remove console
output of ball's spin; it is output in the window's title.