change plain white shapes to images
change ball to sprite to display the entire (extended) texture; the
texture extends past the circle boundaries. change size of paddle
rectangles so that they reach the edge of the window and texture them.
use separate collision box for ball. only collide with side of paddle if
near the front.