add Enemy-Player Collision
enemy colliding with player now also ends the game.
the reason the game is ended is stored in a temporary enum called
"progression" during the update function. the state is progressed based
on this enum. currently, both enum values (other than none - no
progression) affect the state in an identical way.
Player stores its own size. Graphics uses this size to draw the player.
Update uses this size to create a bounding box to use for collisions.