Add Player Death State
adds state for player's death which occurs whenever an enemy bullet
collides with the player. the number of lives are tracked and, on death,
are reduced. currently, the death state only shows a message and waits
for 3 seconds before returning back to gameplay (unless there are no
more lives left in which case the game state switches to "over").

the bullet that collided with the player is removed instantaneously.

during the death state "delay", the bullets are continued to be updated,
which allows them to continue "falling"; this clear them from the scene.
however, the rest of the game is "frozen"/"paused" during the death
delay.

added non-const iterators to Enemies so that the enemy can be modified
(so that enemy's bullet can be removed during the loop).