Dynamic Chance of Enemy Fire
make chance of enemy fire dynamic that now changes based on number of
enemies that are still alive. the more enemies alive, the less likely to
fire. when all enemies are alive, this chance is similar to the previous
one but when enemies are fewer, the chance of them firing increases
considerably. this means that there should still be enemies firing when
there are not many left.