Reduce Enemy Speed Increases
reduce the amount that enemies' speeds increase horizontally and
vertically when one of them is killed to make it slightly easier. this
is because since adding the covers, it's much harder to kill all enemies
in time. this adjustment makes the difficulty closer to the original one
before adding the covers.