Fix High Score
fix high score not being updated correctly.

during the update, the score was adjusted 'after' checking if it was higher than the high score so it could be higher without updating the high score. when the level was "won", the final score was not checked against the high score. this could mean that the high score wouldn't be as high as the final score.

this has been fixed so that the high score is checked 'after' the score has been updated.